Extract Remember\Token to Token\SplitToken
The split-token (selector : validator) scheme is transport-agnostic: the
same primitive backs "remember me" cookies and the opaque API tokens
planned in #110. Move it out of Remember/ so both features share one
audited implementation, and rename makeCookieValue()/parseCookieValue()
to makeValue()/parseValue() now that the value is not always a cookie.

No behaviour change. Done ahead of the rest of #110 because Remember/ is
still unreleased, so the move costs nothing today; after 6.x is tagged it
would require a deprecation cycle.
